--- src/_tests_/storage.test.js
@@ -1,30 +1,0 @@
-const addTodo = require('../add_remove');
-
-const localStorageMock = (() => {
-  let store = {};
-  return {
-    getItem(key) {
-      return store[key];
-    },
-    setItem(key, value) {
-      store[key] = value.toString();
-    },
-    clear() {
-      store = {};
-    },
-    removeItem(key) {
-      delete store[key];
-    }
-  };
-})();
-Object.defineProperty(window, 'localStorage', {
-  value: localStorageMock
-});
-
-test('Add item to the list', () => {
-  const task = 'test';
-  addTodo(task);
-  expect(localStorage.getItem('todo')).toContain(task);
-}
-);
-
